+++ .gitignore
@@ -1,0 +1,3 @@
+models/
+data/
+__pycache__/
