--- train.py
@@ -1,344 +1,0 @@
-from dataclasses import dataclass
-import inspect
-import torch
-import torch.nn as nn
-from torch.nn import functional as F
-import math
-import random
-
-class CausalSelfAttention(nn.Module):
-    def __init__(self, config):
-        super().__init__()
-        assert config.n_embd % config.n_head == 0
-        self.c_attn = nn.Linear(config.n_embd, config.n_embd * 3)
-        self.c_proj = nn.Linear(config.n_embd, config.n_embd)
-        
-        self.n_head = config.n_head
-        self.n_embd = config.n_embd
-        
-        self.register_buffer("bias", torch.tril(torch.ones(config.block_size, config.block_size))
-                             .view(1, 1, config.block_size, config.block_size))
-    
-    def forward(self, x):
-        B, T, C = x.size()
-        qkv = self.c_attn(x)
-        q, k, v = qkv.split(self.n_embd, dim=2)
-        k = k.view(B, T, self.n_head, C // self.n_head).transpose(1, 2)
-        q = q.view(B, T, self.n_head, C // self.n_head).transpose(1, 2)
-        v = v.view(B, T, self.n_head, C // self.n_head).transpose(1, 2)
-        
-        y = F.scaled_dot_product_attention(q, k, v, is_causal=True) # flash attention
-        
-        y = y.transpose(1, 2).contiguous().view(B, T, C)
-        y = self.c_proj(y)
-        return y
-
-class MLP(nn.Module):
-    def __init__(self, config) -> None:
-        super().__init__()
-        self.c_fc = nn.Linear(config.n_embd, config.n_embd * 4)
-        self.gelu = nn.GELU(approximate='tanh')
-        self.c_proj = nn.Linear(config.n_embd * 4, config.n_embd)
-        self.c_proj.NANOGPT_SCALE_INIT = 1
-        
-    def forward(self, x):
-        x = self.c_fc(x)
-        x = self.gelu(x)
-        x = self.c_proj(x)
-        return x
-
-class Block(nn.Module):
-    def __init__(self, config):
-        super().__init__()
-        self.ln_1 = nn.LayerNorm(config.n_embd)
-        self.attn = CausalSelfAttention(config)
-        self.ln_2 = nn.LayerNorm(config.n_embd)
-        self.mlp = MLP(config)
-        
-    def forward(self, x):
-        x = x + self.attn(self.ln_1(x))
-        x = x + self.mlp(self.ln_2(x))
-        return x
-
-@dataclass
-class GPTConfig:
-    block_size: int = 1024
-    vocab_size: int = 10000
-    n_layer: int = 12
-    n_head: int = 12
-    n_embd: int = 768
-
-class GPT(nn.Module):
-    
-    def __init__(self, config) -> None:
-        super().__init__()
-        self.config = config
-        
-        self.transformer = nn.ModuleDict(dict(
-            wte = nn.Embedding(config.vocab_size, config.n_embd),
-            wpe = nn.Embedding(config.block_size, config.n_embd),
-            h = nn.ModuleList([Block(config) for _ in range(config.n_layer)]),
-            ln_f = nn.LayerNorm(config.n_embd),
-        ))
-        self.lm_head = nn.Linear(config.n_embd, config.vocab_size, bias=False)
-        self.transformer.wte.weight = self.lm_head.weight
-        
-        self.apply(self._init_weights)
-    
-    def _init_weights(self, module):
-        if isinstance(module, nn.Linear):
-            std = 0.02
-            if hasattr(module, 'NANOGPT_SCALE_INIT'):
-                std *= (2 * self.config.n_layer) ** -0.5
-            torch.nn.init.normal_(module.weight, mean=0.0, std=0.02)
-            if module.bias is not None:
-                torch.nn.init.zeros_(module.bias)
-        elif isinstance(module, nn.Embedding):
-            torch.nn.init.normal_(module.weight, mean=0.0, std=0.02)
-        
-    def forward(self, idx, targets=None):
-        # idx is of shape (B, T)
-        B, T = idx.size()
-        assert T <= self.config.block_size, f"Cannot forward sequence of length {T}, block size is only {self.config.block_size}"
-        # forward the token and posisition embeddings
-        pos = torch.arange(0, T, dtype=torch.long, device=idx.device) # shape (T)
-        pos_emb = self.transformer.wpe(pos) # position embeddings of shape (T, n_embd)
-        tok_emb = self.transformer.wte(idx) # token embeddings of shape (B, T, n_embd)
-        x = tok_emb + pos_emb
-        # forward the blocks of the transformer
-        for block in self.transformer.h:
-            x = block(x)
-        # forward the final layernorm and the classifier
-        x = self.transformer.ln_f(x)
-        logits = self.lm_head(x) # (B, T, vocab_size)
-        loss = None
-        if targets is not None:
-            loss = F.cross_entropy(logits.view(-1, logits.size(-1)), targets.view(-1))
-        return logits, loss
-    
-    @classmethod
-    def from_pretrained(cls, model_type):
-        """Loads pretrained GPT-2 model weights from huggingface"""
-        assert model_type in {'gpt2', 'gpt2-medium', 'gpt2-large', 'gpt2-xl'}
-        from transformers import GPT2LMHeadModel
-        print("loading weights from pretrained gpt: %s" % model_type)
-
-        # n_layer, n_head and n_embd are determined from model_type
-        config_args = {
-            'gpt2':         dict(n_layer=12, n_head=12, n_embd=768),  # 124M params
-            'gpt2-medium':  dict(n_layer=24, n_head=16, n_embd=1024), # 350M params
-            'gpt2-large':   dict(n_layer=36, n_head=20, n_embd=1280), # 774M params
-            'gpt2-xl':      dict(n_layer=48, n_head=25, n_embd=1600), # 1558M params
-        }[model_type]
-        config_args['vocab_size'] = 50257 # always 50257 for GPT model checkpoints
-        config_args['block_size'] = 1024 # always 1024 for GPT model checkpoints
-        # create a from-scratch initialized minGPT model
-        config = GPTConfig(**config_args)
-        model = GPT(config)
-        sd = model.state_dict()
-        sd_keys = sd.keys()
-        sd_keys = [k for k in sd_keys if not k.endswith('.attn.bias')] # discard this mask / buffer, not a param
-
-        # init a huggingface/transformers model
-        model_hf = GPT2LMHeadModel.from_pretrained(model_type)
-        sd_hf = model_hf.state_dict()
-
-        # copy while ensuring all of the parameters are aligned and match in names and shapes
-        sd_keys_hf = sd_hf.keys()
-        sd_keys_hf = [k for k in sd_keys_hf if not k.endswith('.attn.masked_bias')] # ignore these, just a buffer
-        sd_keys_hf = [k for k in sd_keys_hf if not k.endswith('.attn.bias')] # same, just the mask (buffer)
-        transposed = ['attn.c_attn.weight', 'attn.c_proj.weight', 'mlp.c_fc.weight', 'mlp.c_proj.weight']
-        # basically the openai checkpoints use a "Conv1D" module, but we only want to use a vanilla Linear
-        # this means that we have to transpose these weights when we import them
-        assert len(sd_keys_hf) == len(sd_keys), f"mismatched keys: {len(sd_keys_hf)} != {len(sd_keys)}"
-        for k in sd_keys_hf:
-            if any(k.endswith(w) for w in transposed):
-                # special treatment for the Conv1D weights we need to transpose
-                assert sd_hf[k].shape[::-1] == sd[k].shape
-                with torch.no_grad():
-                    sd[k].copy_(sd_hf[k].t())
-            else:
-                # vanilla copy over the other parameters
-                assert sd_hf[k].shape == sd[k].shape
-                with torch.no_grad():
-                    sd[k].copy_(sd_hf[k])
-
-        return model
-    
-    def configure_optimizers(self, weight_decay, learning_rate, device):
-        # start with all of the candidate parameters (that require grad)
-        param_dict = {pn: p for pn, p in self.named_parameters()}
-        param_dict = {pn: p for pn, p in param_dict.items() if p.requires_grad}
-        # create optim groups. Any parameters that is 2D will be weight decayed, otherwise no.
-        # i.e. all weight tensors in matmuls + embeddings decay, all biases and layernorms don't.
-        decay_params = [p for n, p in param_dict.items() if p.dim() >= 2]
-        nodecay_params = [p for n, p in param_dict.items() if p.dim() < 2]
-        optim_groups = [
-            {'params': decay_params, 'weight_decay': weight_decay},
-            {'params': nodecay_params, 'weight_decay': 0.0}
-        ]
-        num_decay_params = sum(p.numel() for p in decay_params)
-        num_nodecay_params = sum(p.numel() for p in nodecay_params)
-        # if master_process:
-        print(f"num decayed parameter tensors: {len(decay_params)}, with {num_decay_params:,} parameters")
-        print(f"num non-decayed parameter tensors: {len(nodecay_params)}, with {num_nodecay_params:,} parameters")
-        # Create AdamW optimizer and use the fused version if it is available
-        fused_available = 'fused' in inspect.signature(torch.optim.AdamW).parameters
-        use_fused = fused_available and device == "cuda"
-        # if master_process:
-        print(f"using fused AdamW: {use_fused}")
-        optimizer = torch.optim.AdamW(optim_groups, lr=learning_rate, betas=(0.9, 0.95), eps=1e-8, fused=use_fused)
-        return optimizer
-    
-# ------------------------------------------------------------------------------
-from tokenizer import RegexTokenizer
-
-class DataLoaderLite:
-    def __init__(self, B, T, file=None):
-        self.B = B
-        self.T = T
-        self.enc = RegexTokenizer("merges.json", "vocab.json")
-        
-        if file is not None:
-            import json
-            with open(file, 'r', encoding='utf-8') as f:
-                tokens = json.load(f)
-        else:
-            with open('output-merged.txt', 'r', encoding='utf-8') as f:
-                text = f.read()
-            tokens = self.enc.encode(text)
-        
-        self.tokens = torch.tensor(tokens)
-        print(f'loaded {len(tokens)} tokens')
-        print(f'1 epoch = {len(tokens) // (B * T)} batches')
-        
-        self.current_position = 0
-    
-    def next_batch(self):
-        B, T = self.B, self.T
-        buf = self.tokens[self.current_position:self.current_position + B*T + 1]
-        x = buf[:-1].view(B, T) # inputs
-        y = buf[1:].view(B, T) # targets
-        
-        # move position forwards
-        self.current_position += B*T
-        # reset if out of bounds
-        if self.current_position + B*T + 1 >= len(self.tokens):
-            self.current_position = 0
-        return x, y
-
-# ------------------------------------------------------------------------------
-class FriendLLM:
-    def __init__(self) -> None:
-        self.device = 'cpu'
-        if torch.cuda.is_available():
-            self.device = 'cuda'
-        elif hasattr(torch.backends, 'mps') and torch.backends.mps.is_available():
-            self.device = 'mps'
-        print('using device:', self.device)
-        
-        # get logits
-        self.model = GPT(GPTConfig(vocab_size=10112))
-        self.model.to(self.device)
-        
-        self.enc = RegexTokenizer("merges.json", "vocab.json", "bad.json")
-        
-        print(self.enc.banned_tokens)
-        
-    def train_model(self, steps):
-        import time
-
-        train_loader = DataLoaderLite(B=16, T=512, file='bytes2.json')
-        torch.set_float32_matmul_precision('high')
-
-        # torch.compile slows down training for some reason???
-        # model = torch.compile(model, backend="aot_eager")
-
-        max_lr = 6e-4
-        min_lr = max_lr * 0.1
-        warmup_steps = int(steps * 0.025)
-        max_steps = steps
-
-        def get_lr(it):
-            # warmup
-            if it < warmup_steps:
-                return max_lr * (it + 1) / warmup_steps
-            # min learning rate
-            if it > max_steps:
-                return min_lr
-            # decay
-            decay_ratio = (it - warmup_steps) / (max_steps - warmup_steps)
-            assert 0 <= decay_ratio <= 1
-            coeff = 0.5 * (1 + math.cos(math.pi * decay_ratio))
-            return min_lr + coeff * (max_lr - min_lr)
-
-        optimizer = self.model.configure_optimizers(weight_decay=0.1, learning_rate=6e-4, device=self.device)
-
-        for step in range(max_steps):
-            t0 = time.time()
-            x, y = train_loader.next_batch()
-            x, y = x.to(self.device), y.to(self.device)
-            optimizer.zero_grad()
-            with torch.autocast(device_type=self.device, dtype=torch.bfloat16):
-                logits, loss = self.model(x, y)
-            loss.backward()
-            norm = torch.nn.utils.clip_grad_norm_(self.model.parameters(), 1.0)
-            
-            lr = get_lr(step)
-            for param_group in optimizer.param_groups:
-                param_group['lr'] = lr
-            optimizer.step()
-            
-            torch.cuda.synchronize()
-            t1 = time.time()
-            print(f"step {step:4d}, loss: {loss.item()}, lr: {lr:.4e}, norm: {norm:.4f} dt: {((t1-t0)*1000):.4f}ms")
-
-        torch.save(self.model.state_dict(), f'model_{max_steps}.pth')
-
-    def load_model(self, file):
-        self.model.load_state_dict(torch.load(file, weights_only=True))
-        
-    def prompt(self, prompt):
-        message_count = 0
-        message_count_hard_cap = 60
-
-        tokens = self.enc.encode(prompt)
-        tokens = torch.tensor(tokens).unsqueeze(0)
-        x = tokens.to(self.device)
-        
-        def get_stop_prob(it):
-            if it > message_count_hard_cap:
-                return 0
-
-            decay_ratio = it / message_count_hard_cap
-            assert 0 <= decay_ratio <= 1
-            return 0.5 * (1 + math.cos(math.pi * decay_ratio))
-
-        # torch.manual_seed(42)
-        # torch.cuda.manual_seed(42)
-        while message_count < message_count_hard_cap:
-            with torch.no_grad():
-                logits, _ = self.model(x)
-                logits = logits[:, -1, :]
-                probs = F.softmax(logits, dim=-1)
-                topk_probs, topk_indices = torch.topk(probs, 50, dim=-1)
-                ix = torch.multinomial(topk_probs, 1)
-                xcol = torch.gather(topk_indices, -1, ix)
-                if xcol.item() in self.enc.banned_tokens:
-                    print('bad token:', xcol.item())
-                    continue
-                
-                x = torch.cat((x, xcol), dim=1)
-                
-                if xcol.item() in self.enc.eom_tokens:
-                    if(random.random() > get_stop_prob(message_count)):
-                        break
-                    message_count += 1
-
-        tokens = x[0, :].tolist()
-        decoded = self.enc.decode(tokens)
-        print(decoded)
-
-bot = FriendLLM()
-bot.load_model('model_8000.pth')
-bot.prompt("gqds:")
